api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi==23.1.0",
    "bcrypt>=5.0.0",
    "beanie==1.30.0",
    "email-validator==2.2.0",
    "fastapi==0.115.0",
    "motor==3.6.0",
    "passlib>=1.7.4",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "pymongo==4.9.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "uvicorn[standard]==0.32.0",
]
